feature: Prevent deletion of categories with associated transactions and enhance error handling.

+++ app/domain/use_case/financial/category/destroy_category_use_case.rb
@@ -6,6 +6,11 @@ class UseCase::Financial::Category::DestroyCategoryUseCase
   # @return [Financial::Category]
   def call(user:, id:)
     category = user.categories.find(id)
+
+    if category.transactions.exists?
+      raise ArgumentError, "Cannot delete category with associated transactions"
+    end
+
     category.destroy!
     category
   end
